Update workflows so that Rockstar version is passed between actions
Previously, the Linux build would patch the version number to 2.0.{{ run_number }} and then push this back to the repo - but I'd incorrectly assumed that the `checkout` actions in the called workflows would pull the latest. They actually pull the commit that triggered the parent workflow, so they'll never see the correct version number.