Only update pronoun subject if assignment target is NOT an array element
Previously, this code would cause an infinite loop:

```
for x in 3
n = 0
array at it = 0
```

First time through, 'array' is undefined at the point we assign 'it' -  so we assign `array[mysterious] = 0`
Second time through, `array` exists,  so `array at it = 0` assigns `array[array] = 0` - and updates `it` to point to `array[array]` but **does not actually evaluate it**

But once you've got an array that uses itself as an index, any attempt to assign, read, write, debug, etc. that array causes an infinite loop.

This fix modifies the code so that the pronoun subject is only updated AFTER the assignment operation has completed, so any pronouns used in array index expressions will retain the values they had when the assignment started and won't change value halfway through.

Fixes: #368